setup

+++ src/cuda/gemm_tensorcores/main.cu
@@ -619,7 +619,7 @@ int main(int argc, char** argv) {
 		call_mxm<float, float, half, float>(host_matrix_a, host_matrix_b, log_obj);
 	}
 	if (log_obj.precision == "double") {
-		call_mxm<double, double, half, float>(host_matrix_a, host_matrix_b, log_obj);
+		call_mxm<double, double, half, double>(host_matrix_a, host_matrix_b, log_obj);
 	}
 	if (log_obj.precision == "DMR") {
 		call_mxm<double, double, half, float>(host_matrix_a, host_matrix_b, log_obj);
